[fix] Move results_xxx to utils from API
